fix(plc4go/codgen): fixed array and string problems

+++ plc4go/internal/plc4go/ads/readwrite/model/AdsStaticHelper.go
@@ -43,5 +43,5 @@ func StaticHelperSerializeAmsString(io utils.WriteBuffer, value values.PlcValue,
 	case "UTF-16":
 		multiplier = 16
 	}
-	return io.WriteString("", uint8(stringLength*multiplier), encoding, value.GetString())
+	return io.WriteString("", uint32(stringLength*multiplier), encoding, value.GetString())
 }
