+++ plc4go/internal/plc4go/spi/utils/WriteBufferBoxBased.go
@@ -177,7 +177,7 @@ func (b *boxedWriteBuffer) WriteBigFloat(logicalName string, bitLength uint8, va
 	return nil
 }
 
-func (b *boxedWriteBuffer) WriteString(logicalName string, bitLength uint8, _ string, value string, writerArgs ...WithWriterArgs) error {
+func (b *boxedWriteBuffer) WriteString(logicalName string, bitLength uint32, _ string, value string, writerArgs ...WithWriterArgs) error {
 	additionalStringRepresentation := b.extractAdditionalStringRepresentation(upcastWriterArgs(writerArgs...)...)
 	b.PushBack(BoxString(logicalName, fmt.Sprintf("%#0*x %s%s", bitLength/4, value, value, additionalStringRepresentation), 0))
 	return nil
